pruebas

+++ src/main/java/com/unla/grupo12OO22024/models/VentaModel.java
@@ -19,7 +19,7 @@ public class VentaModel {
 
 	@Id
 	@Setter(AccessLevel.PROTECTED)
-	private int id_venta;
+	private long id_venta;
 	
 	private Producto producto;
